requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-chroma==0.1.2
langchain-groq==0.1.9
python-dotenv==1.0.1
sentence-transformers==3.0.1
chromadb>=0.5.0,<0.5.4